requirements.txt
asgiref==3.2.3
asyncore-wsgi==0.0.9
bottle==0.12.17
# Editable install with no version control (Django==2.2.7)
-e /home/farid/Sites/virtualenvs/django227/lib/python3.6/site-packages
django-debug-toolbar==2.0
django-filter==2.2.0
Faker==2.0.3
mysql-connector-python-rf==2.2.2
mysqlclient==1.4.5
pkg-resources==0.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
web-pdb==1.5.1
